Audio: Set up service
In order for audio to play properly (and to get around Safari's audio
restrictions), I've set up a service that will handle all state
management of audio. The <Audio /> component will subscribe to this
service and conditionally add/remove/update the audio HTML element.

As long as we use the same <audio> element, we can get around Safari's
restrictions on playing audio.